Updated crud opertions to live on models now
